refactor preprocess to use EagerModelBase
[ghstack-poisoned]
